Added the following files: Book.java, HomePage.java, activity_home_page.xml, MyBooks.java,activity_my_books.xml,CustomBookAdapter.java,AddBookFragment.java, add_book_fragment_layout.xml,bookcontent.xml
When the user successfully logs in, HomePage activity opens up where a button for "MY BOOKS" is displayed. When the users clicks the button, a new activity MyBooks opens up where the user can see each a list of all the books that they own with their title,author,ISBN and status. In this activity, a floating action button is added as well that opens up a fragment which gives the functionality to add a new book.
When the user adds a new book, it is visible in the list. Implemented real-time updates using firebase so the changes made in the server are seen in the app simultaneously.

User Stories Done: 01.01.01
User Stories partially done: 01.03.01, 01.04.01.