Fixed Bugs & Input
Fixed bugs in some activites that were causing the app to crash. The
main one fixed is Delete button in a book picture. If the book was not
already there then delete buttton caused an error that crashed the app.
Some code formatting done to make code look more pleasing or
professional. Still have to fix some variable names that are so
descriptive about what they do. Fixed Date of Birth input validation.
Now user can select a date from the dialog, so there will be no errors
in that :).